Exp batch: fix issue with Flush()
Since Flush retrives pending draws and issues new-ones, as seen from
outside of the package, it does not flush anything at all and does not
work as expected. This commit renames it to an unexported flush() and
exports a new Flush() function that calls flush() once then calls it
again it there are any pending draw commands.