Revert a11eed6: texture bleeding workaround should not be hardcoded.
The method implemented worked well with the demo material but was not
bulletproof nor ideal for all users. There are so many ways around this
that it should be left to the user to decide how to do it.

The workaround has been moved to the demo code as a custom
grog.Drawable.